FDG-6461 update tests
