+++ src/layouts/explainer/explainer-components/afg-components/afg-topic-card/afg-topic-card.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Grid } from '@material-ui/core';
 import { ChartPlaceholder } from '../../../explainer-helpers/national-deficit/national-deficit-helper';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faArrowRightLong } from '@fortawesome/free-solid-svg-icons';
