--- src/layouts/explainer/explainer-components/afg-components/afg-topic-card/afg-topic-card.spec.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import AfgTopicCard from './afg-topic-card';
-import { render } from '@testing-library/react';
-
-describe('Topic Section Component', () => {
-  it('renders the component', () => {
-    const { getByTestId } = render(
-      <AfgTopicCard
-        heading="The amount by which spending exceeds revenue, $X.X in YYYY, is referred to as deficit spending."
-        body="A budget deficit occurs when the money spent exceeds the money collected for a given period."
-        linkUrl="/national-deficit"
-        linkText="Learn more about national deficit"
-        image="/topics-section-images/homepage_deficit_1200x630.png"
-        imageAltText=""
-        id="national-deficit"
-      />
-    );
-    expect(getByTestId('topic-section')).toBeInTheDocument();
-  });
-});
