--- src/layouts/explainer/explainer-components/afg-components/compare-section/compare-section.spec.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import CompareSection from './compare-section';
-import { render } from '@testing-library/react';
-
-describe('Compare Section Component', () => {
-  it('renders the component', () => {
-    const { getByTestId } = render(<CompareSection currentFiscalYear={2022} />);
-    expect(getByTestId('compare-section')).toBeInTheDocument();
-  });
-});
